+++ components/core/blocks/Header/Header.vue
@@ -5,59 +5,8 @@
       :class="{ 'is-visible': navVisible }"
     >
       <div class="container px15">
-        <div class="row between-xs middle-xs" v-if="!isCheckoutPage || isThankYouPage">
-          <div class="col-md-4 col-xs-2 middle-xs">
-            <div>
-              <hamburger-icon class="p15 icon bg-cl-secondary pointer" />
-            </div>
-          </div>
-          <div class="col-xs-2 visible-xs">
-            <search-icon class="p15 icon pointer" />
-          </div>
-          <div class="col-md-4 col-xs-4 center-xs pt5">
-            <div>
-              <logo width="auto" height="41px" />
-            </div>
-          </div>
-          <div class="col-xs-2 visible-xs">
-            <wishlist-icon class="p15 icon pointer" />
-          </div>
-          <div class="col-md-4 col-xs-2 end-xs">
-            <div class="inline-flex right-icons">
-              <search-icon class="p15 icon hidden-xs pointer" />
-              <wishlist-icon class="p15 icon hidden-xs pointer" />
-              <compare-icon class="p15 icon hidden-xs pointer" />
-              <microcart-icon class="p15 icon pointer" />
-              <account-icon class="p15 icon hidden-xs pointer" />
-            </div>
-          </div>
-        </div>
-        <div class="row between-xs middle-xs px15 py5" v-if="isCheckoutPage && !isThankYouPage">
-          <div class="col-xs-5 col-md-3 middle-xs">
-            <div>
-              <router-link
-                :to="localizedRoute('/')"
-                class="cl-tertiary links"
-              >
-                {{ $t('Return to shopping') }}
-              </router-link>
-            </div>
-          </div>
-          <div class="col-xs-2 col-md-6 center-xs">
-            <logo width="auto" height="41px" />
-          </div>
-          <div class="col-xs-5 col-md-3 end-xs">
-            <div>
-              <a
-                v-if="!currentUser"
-                href="#"
-                @click.prevent="gotoAccount"
-                class="cl-tertiary links"
-              >{{ $t('Login to your account') }}</a>
-              <span v-else>{{ $t('You are logged in as {firstname}', currentUser) }}</span>
-            </div>
-          </div>
-        </div>
+        <CommonHeaderItems v-if="!isCheckoutPage || isThankYouPage" />
+        <CheckoutHeaderItems v-if="isCheckoutPage && !isThankYouPage" />
       </div>
     </header>
     <div class="header-placeholder" />
@@ -65,26 +14,13 @@
 </template>
 
 <script>
-import { mapState } from 'vuex'
 import CurrentPage from 'theme/mixins/currentPage'
-import AccountIcon from 'theme/components/core/blocks/Header/AccountIcon'
-import CompareIcon from 'theme/components/core/blocks/Header/CompareIcon'
-import HamburgerIcon from 'theme/components/core/blocks/Header/HamburgerIcon'
-import Logo from 'theme/components/core/Logo'
-import MicrocartIcon from 'theme/components/core/blocks/Header/MicrocartIcon'
-import SearchIcon from 'theme/components/core/blocks/Header/SearchIcon'
-import WishlistIcon from 'theme/components/core/blocks/Header/WishlistIcon'
 
 export default {
   name: 'Header',
   components: {
-    AccountIcon,
-    CompareIcon,
-    HamburgerIcon,
-    Logo,
-    MicrocartIcon,
-    SearchIcon,
-    WishlistIcon
+    CommonHeaderItems: () => import('theme/components/core/blocks/Header/CommonHeaderItems'),
+    CheckoutHeaderItems: () => import('theme/components/core/blocks/Header/CheckoutHeaderItems')
   },
   mixins: [CurrentPage],
   data () {
@@ -96,15 +32,9 @@ export default {
       navbarHeight: 54
     }
   },
-  computed: {
-    ...mapState({
-      isOpenLogin: state => state.ui.signUp,
-      currentUser: state => state.user.current
-    }),
+  computed: {   
     isThankYouPage () {
       return this.$store.state.checkout.isThankYouPage
-        ? this.$store.state.checkout.isThankYouPage
-        : false
     }
   },
   beforeMount () {
@@ -124,9 +54,6 @@ export default {
     }, 250)
   },
   methods: {
-    gotoAccount () {
-      this.$bus.$emit('modal-toggle', 'modal-signup')
-    },
     hasScrolled () {
       this.scrollTop = window.scrollY
       if (