+++ layouts/Default.vue
@@ -30,9 +30,7 @@
         <main-footer />
       </lazy-hydrate>
 
-      <lazy-hydrate when-visible>
-        <notification />
-      </lazy-hydrate>
+      <notification v-if="notifications"/>
 
       <lazy-hydrate when-idle>
         <sign-up />
@@ -57,12 +55,9 @@ import { mapState } from 'vuex'
 import AsyncSidebar from 'theme/components/theme/blocks/AsyncSidebar/AsyncSidebar.vue'
 import MainHeader from 'theme/components/core/blocks/Header/Header.vue'
 import MainFooter from 'theme/components/core/blocks/Footer/Footer.vue'
-import Overlay from 'theme/components/core/Overlay.vue'
 import Loader from 'theme/components/core/Loader.vue'
-import Notification from 'theme/components/core/Notification.vue'
 import SignUp from 'theme/components/core/blocks/Auth/SignUp.vue'
 import CookieNotification from 'theme/components/core/CookieNotification.vue'
-import OfflineBadge from 'theme/components/core/OfflineBadge.vue'
 import { isServer } from '@vue-storefront/core/helpers'
 import Head from 'theme/head'
 import config from 'config'
@@ -72,7 +67,10 @@ import LazyHydrate from 'vue-lazy-hydration'
 const SidebarMenu = () => import(/* webpackPreload: true */ /* webpackChunkName: "vsf-sidebar-menu" */ 'theme/components/core/blocks/SidebarMenu/SidebarMenu.vue')
 const Microcart = () => import(/* webpackPreload: true */ /* webpackChunkName: "vsf-microcart" */ 'theme/components/core/blocks/Microcart/Microcart.vue')
 const Wishlist = () => import(/* webpackPreload: true */ /* webpackChunkName: "vsf-wishlist" */ 'theme/components/core/blocks/Wishlist/Wishlist.vue')
+const Notification = () => import(/* webpackPreload: true */ /* webpackChunkName: "vsf-notification" */ 'theme/components/core/Notification.vue')
+const OfflineBadge = () => import(/* webpackPreload: true */ /* webpackChunkName: "vsf-offline-badge" */ 'theme/components/core/OfflineBadge.vue')
 const SearchPanel = () => import(/* webpackChunkName: "vsf-search-panel" */ 'theme/components/core/blocks/SearchPanel/SearchPanel.vue')
+const Overlay = () => import(/* webpackChunkName: "vsf-overlay" */ 'theme/components/core/Overlay.vue')
 const OrderConfirmation = () => import(/* webpackChunkName: "vsf-order-confirmation" */ 'theme/components/core/blocks/Checkout/OrderConfirmation.vue')
 
 export default {
@@ -93,7 +91,10 @@ export default {
       isSidebarOpen: state => state.ui.sidebar,
       isMicrocartOpen: state => state.ui.microcart,
       isWishlistOpen: state => state.ui.wishlist
-    })
+    }),
+    notifications () {
+      return this.$store.getters['notification/notifications']
+    }
   },
   methods: {
     onOrderConfirmation (payload) {